# openapi.yaml
openapi: 3.0.3
info:
  title: Invoice & Payment Service API
  description: |
    Dodo Payments take-home — Invoice & Payment Service.

    - **Auth:** `Authorization: Bearer <api_key>` on all routes except `/health`
    - **Business key:** printed on `docker compose up` (admin — customers, api-keys, webhooks)
    - **Customer key:** returned once per `POST /customers` (`api_key` in response)
    - **Money:** integer cents (`int64`); server computes invoice totals
    - **Pay:** requires `Idempotency-Key` header; 10s PSP timeout
    - **Mock PSP:** http://localhost:9090 (host) or http://mock-psp:9090 (docker network)
    - **Testing:** see README.md
  version: 1.0.0
  contact:
    name: Invoice Service
servers:
  - url: http://localhost:8080
    description: Local API (docker compose)
  - url: http://localhost:9090
    description: Mock PSP (from host / curl)
...
paths:
...
  /webhook-endpoints:
    post:
      tags: [Webhooks]
      summary: Register webhook endpoint
      operationId: createWebhookEndpoint
      description: Requires business Bearer token. Secret returned only at creation time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook endpoint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpointCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Webhooks]
      summary: List webhook endpoints
      operationId: listWebhookEndpoints
      description: Requires business Bearer token.
      responses:
        '200':
          description: Webhook endpoints (no secret)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpointListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Business key (docker logs) or customer key (POST /customers response)
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: bad_request
            message:
              type: string
              example: Invalid JSON body
...
    CreateWebhookRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: https://webhook.site/your-uuid
    WebhookEndpointCreated:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        secret:
          type: string
          description: Shown only once
        created_at:
          type: string
          format: date-time
    WebhookEndpointListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        business_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    WebhookEndpointListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEndpointListItem'
        count:
          type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
security:
  - BearerAuth: []